+++ resources/js/components/FormTabs.vue
@@ -14,7 +14,7 @@
       </div>
     </div>
 
-    <div class="relationship-tabs-panel card w-full overflow-hidden">
+    <div class="relationship-tabs-panel card w-full">
       <div class="flex flex-row">
         <div
           class="py-5 px-8 border-b-2 focus:outline-none tab cursor-pointer"
@@ -33,6 +33,7 @@
       >
         <div :class="{'px-6 py-3':!tab.listable}">
           <component
+            ref="fields"
             v-for="(field, index) in tab.fields"
             :class="{'remove-bottom-border': index == tab.fields.length - 1}"
             :key="'tab-' + index"
@@ -46,6 +47,7 @@
             :via-resource-id="viaResourceId"
             :via-relationship="viaRelationship"
             @actionExecuted="actionExecuted"
+            :show-help-text="field.helpText != null"
           />
         </div>
       </div>
@@ -118,7 +120,9 @@ export default {
       tabs[field.tab].fields.push(field);
     });
     this.tabs = tabs;
-    this.handleTabClick(tabs[Object.keys(tabs)[0]]);
+    this.handleTabClick({
+        name: this.$route.query.tab || tabs[Object.keys(tabs)[0]].name
+    });
   },
   methods: {
     /**
@@ -136,7 +140,16 @@ export default {
       this.$emit("actionExecuted");
     },
     handleTabClick(tab, event) {
+      let cur = this.$router.currentRoute.query;
       this.activeTab = tab.name;
+
+      if(!cur || cur.tab != tab.name) {
+        this.$router.replace({query: { tab: tab.name }});
+      }
+
+      // When code fields are not visible initially they are not loaded
+      // See https://stackoverflow.com/questions/8349571/codemirror-editor-is-not-loading-content-until-clicked
+      setTimeout(this.refreshCodeFields, 1);
     },
 
     tabHasErrors(tab) {
@@ -152,7 +165,13 @@ export default {
       tab.hasErrors = hasErrors;
 
       return hasErrors;
-    }
+    },
+
+    refreshCodeFields() {
+      this.$refs.fields
+        .filter(field => 'codemirror' in field)
+        .forEach(field => field.codemirror.refresh());
+    },
   }
 };
 </script>